+++ src/main/java/com/github/mwttg/sjge/utilities/ReadTextFile.java
@@ -1,5 +1,8 @@
 package com.github.mwttg.sjge.utilities;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
@@ -12,6 +15,8 @@ import java.util.stream.Collectors;
  */
 public class ReadTextFile {
 
+    private static final Logger LOG = LoggerFactory.getLogger(ReadTextFile.class);
+
     /**
      * reads a text file from the resource folder
      *
@@ -20,6 +25,7 @@ public class ReadTextFile {
      * @throws IOException if read fails for any reason
      */
     public static Collection<String> fromResources(final String filename) throws IOException {
+        LOG.debug("Try to read file: " + filename);
         final var classLoader = ClassLoader.getSystemClassLoader();
         try (final InputStream inputStream = classLoader.getResourceAsStream(filename)) {
             if (inputStream == null) {